requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.2.2
certifi==2020.6.20
chardet==3.0.4
cheroot==8.4.5
distlib==0.3.1
Django==3.2.8
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
jaraco.functools==3.0.1
more-itertools==8.5.0
Pillow==7.2.0
psycopg2-binary==2.9.2
pymongo==3.11.0
python-dateutil==2.8.2
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.24.0
reverse==0.1.0
simplejson==3.17.2
six==1.16.0
soupsieve==2.0.1
sqlparse==0.4.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.25.10
virtualenv==20.0.31
web.py==0.61
